Implement CA Info REST v2
The code of CAInfoService.java has moved to CAEngine and used for the
current API and the new v2 API.

Introduced a base servlet for all CA APIs and the servlet for the Info
end-point.

No code changes performed.
